add SelectedRouteRepository.kt
