+++ app/src/main/java/com/cornellappdev/transit/models/RouteRepository.kt
@@ -2,7 +2,6 @@ package com.cornellappdev.transit.models
 
 import com.cornellappdev.transit.networking.ApiResponse
 import com.cornellappdev.transit.networking.NetworkApi
-import com.cornellappdev.transit.ui.viewmodels.LocationUIState
 import com.google.android.gms.maps.model.LatLng
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -35,14 +34,6 @@ class RouteRepository @Inject constructor(private val networkApi: NetworkApi) {
     private val _lastRouteFlow: MutableStateFlow<ApiResponse<RouteOptions>> =
         MutableStateFlow(ApiResponse.Pending)
 
-    private val _startPlace: MutableStateFlow<LocationUIState> = MutableStateFlow(
-        LocationUIState.Place("Current Location", LatLng(42.44, -76.50))
-    )
-
-    private val _destPlace: MutableStateFlow<LocationUIState> = MutableStateFlow(
-        LocationUIState.Place("Current Location", LatLng(42.45, -76.51))
-    )
-
     init {
         fetchAllStops()
     }
@@ -61,16 +52,6 @@ class RouteRepository @Inject constructor(private val networkApi: NetworkApi) {
      * A StateFlow holding the last queried location
      */
     val placeFlow = _placeFlow.asStateFlow()
-
-    /**
-     * Pair of the name of the starting location and the coordinates
-     */
-    val startPlace = _startPlace.asStateFlow()
-
-    /**
-     * Pair of the name of the ending location and the coordinates
-     */
-    val destPlace = _destPlace.asStateFlow()
 
     /**
      * Makes a new call to backend for all stops.
@@ -140,20 +121,6 @@ class RouteRepository @Inject constructor(private val networkApi: NetworkApi) {
                 _lastRouteFlow.value = ApiResponse.Error
             }
         }
-    }
-
-    /**
-     * Change start location
-     */
-    fun setStartLocation(location: LocationUIState) {
-        _startPlace.value = location
-    }
-
-    /**
-     * Change end location
-     */
-    fun setEndLocation(location: LocationUIState) {
-        _destPlace.value = location
     }
 
 }
