+++ app/src/main/java/com/cornellappdev/transit/models/SelectedRouteRepository.kt
@@ -1,0 +1,45 @@
+package com.cornellappdev.transit.models
+
+import com.cornellappdev.transit.ui.viewmodels.LocationUIState
+import com.google.android.gms.maps.model.LatLng
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class SelectedRouteRepository @Inject constructor() {
+
+    private val _startPlace: MutableStateFlow<LocationUIState> = MutableStateFlow(
+        LocationUIState.Place("Current Location", LatLng(42.44, -76.50))
+    )
+
+    private val _destPlace: MutableStateFlow<LocationUIState> = MutableStateFlow(
+        LocationUIState.Place("Current Location", LatLng(42.45, -76.51))
+    )
+
+    /**
+     * Pair of the name of the starting location and the coordinates
+     */
+    val startPlace = _startPlace.asStateFlow()
+
+    /**
+     * Pair of the name of the ending location and the coordinates
+     */
+    val destPlace = _destPlace.asStateFlow()
+
+    /**
+     * Change start location
+     */
+    fun setStartLocation(location: LocationUIState) {
+        _startPlace.value = location
+    }
+
+    /**
+     * Change end location
+     */
+    fun setEndLocation(location: LocationUIState) {
+        _destPlace.value = location
+    }
+
+}
